complited assignment

+++ Primes.java
@@ -27,6 +27,6 @@ public class Primes {
                 System.out.println(i);
             }
         }
-        System.out.println("There are " + count + "primes between 2 and " + p +" (" +(double) (count/p)*100 + "% are primes)");
+        System.out.println("There are " + count + "primes between 2 and " + p +" (" +((double)count/(double)p)*100 + "% are primes)");
     }
 }
